Reduce indexers from 3 to 2
